// src/app/contenido/interfaces.ts
export type Language = "es" | "en" | "de";

export type Dioses = "" | "Afrodita" | "Apolo" | "Ares" | "Artemisa" | "Atena" | "Demeter" | "Hefesto" | "Hera" | "Hermes" | "Hestia" | "Poseidon" | "Zeus";